feat: Introduce new Generator infrastructure
The existing generator interface is very specific
to the needs of LLVM backend. It uses such a terms
as 'function', 'node', etc. While those make sense
for one backends, they may have no meaning for other.
As part of effort to bring MLIR into Athena, a new
Generator infrastructure is introduced. It aims to
be flexible and extensible without respect to
underlying backends. LLVM backend can still map
data, that is passed to new generator, to functions
and whatever other structures it needs.

Transition to the new infrastructure will be done
in future patches.

Meta: ATH-13